+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.2.0] - 202?-??-?? (not released yet)
 ### Fixed
+ - Fixed ```ArgumentNullException``` during internal ```DbgDraw``` initialization due to missing a required shader. I've implemented a build pre-processor that automatically adds the required ```Hidden/DbgDraw-Shaded``` shader to the 'Always Included Shaders' list found under 'Project Settings > Graphics'. This fixes issue #6.
  - Fixed ```FindObjectOfType``` deprecated warning when using Unity 2023.1 and newer.
  
 ## [1.1.0] - 2022-01-05